fix: message redux 분리
