+++ front/src/rootReducer.ts
@@ -1,8 +1,10 @@
 import { combineReducers } from '@reduxjs/toolkit';
 import users from './ToolKit/user';
+import message from './ToolKit/messages';
 
 const rootReducer = combineReducers({
   users,
+  message,
 });
 
 export type ReducerType = ReturnType<typeof rootReducer>;
