+++ front/src/views/MessageListView.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { useDispatch, useSelector } from 'react-redux';
-import { listMessage, message } from '../ToolKit/user';
+import { listMessage, message } from '../ToolKit/messages';
 
 function MessageListView() {
   return <div></div>;
